requirements.txt
python-telegram-bot==13.15
pytube==15.0.0
yt-dlp==2023.11.16
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
